+++ lib/opts/inlineCheckbox.ts
@@ -2,25 +2,32 @@ import type { Choice } from '$types';
 import type { PromptOpts } from './prompt.ts';
 
 /**
- * Options for the checkbox prompt.
+ * Options for the inline checkbox prompt.
  */
-export type CheckboxOpts = PromptOpts<unknown[]> & {
+export type InlineCheckboxOpts = PromptOpts<unknown[]> & {
   /**
    * The type of the prompt. This can not be changed but will be used to
    * determine the type of the question.
    */
-  type?: 'checkbox';
+  type?: 'inline-checkbox';
 
   /**
    * A list of choices for the user to select multiple values from.
    */
   choices: Choice[];
 
-  columns?: number;
+  /**
+   * A string to be displayed after the choices.
+   */
+  suffix?: string;
 
   selectedPrefix?: string;
 
   unselectedPrefix?: string;
+
+  finalSelectedPrefix?: string;
+
+  finalUnselectedPrefix?: string;
 
   /**
    * A function that can override the way an unchecked choice is displayed. It