+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.0.3-beta.14] - 2025-09-05
+
+- Added inlineCheckbox prompt.
+- Made the spacing before questions a global preNewline option.
+
 ## [2.0.3-beta.13] - 2025-09-05
 
 - Bug fixes with final prompt output